Update example to use `Init`
